% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/actualizar_package.R
\encoding{UTF-8}
\name{actualizar_package}
\alias{actualizar_package}
\title{Actualizar package}
\usage{
actualizar_package()
}
\value{
Package \code{estyr} actualizado a su versión más reciente.
}
\description{
Función que actualiza el package 'estyr'.
}
\details{
Debido a lo complicado que puede ser desinstalar un package y reinstalarlo. Automatizamos ese procesimiento con esta función.
}
\examples{
actualizar_package()

}
\author{
Esteban Rucán.
}
