% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/crear_presentacion.R
\encoding{UTF-8}
\name{crear_presentacion}
\alias{crear_presentacion}
\title{Crear Presentación de R Xaringan.}
\usage{
crear_presentacion(titulo = "Sin titulo")
}
\arguments{
\item{titulo}{Character. Título de la presentación.}
}
\value{
Plantilla Xaringan en R Markdown.
}
\description{
Crear una presentación de R Markdown en el directorio seleccionado.
}
\details{
Esta función proporciona una forma rápida de crear una presentación en R Markdown.
}
\examples{
\dontrun{crear_presentacion("Analsis Exploratorio")}

}
\author{
Esteban Rucán.
}
