% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/crear_proyecto.R
\encoding{UTF-8}
\name{crear_proyecto}
\alias{crear_proyecto}
\title{Crear Proyecto de RStudio}
\usage{
crear_proyecto(estructura = TRUE)
}
\arguments{
\item{estructura}{Valor lógico. Si es \code{TRUE}, indica que desea crear carpetas para estructurar el proyecto}
}
\value{
Proyecto
}
\description{
Función que crea un proyecto en el directorio indicdo
}
\details{
Función que facilita la creación de un proyecto en RStudio
}
\examples{
dontrun{crear_proyecto()}

}
\author{
Esteban Rucán.
}
