% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/e_theme_estyr.R
\encoding{UTF-8}
\name{e_theme_estyr}
\alias{e_theme_estyr}
\title{Tema para echarts4r}
\usage{
e_theme_estyr(e)
}
\arguments{
\item{e}{Objeto de \code{echarts4r}.}
}
\value{
Tema de echarts4r.
}
\description{
Tema de echarts4r con algunas configuraciones habituales
}
\details{
Est치 funci칩n fue desarrollada con el fin de tener gr치ficos de un solo estilo
}
\examples{
\dontrun{grafico \%>\% e_theme_estyr()}

}
\author{
Esteban Ruc치n.
}
