% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estyr.R
\docType{package}
\name{estyr}
\alias{estyr}
\alias{estyr-package}
\title{Package para facilitar el trabajo en R}
\description{
Este package brinda opciones muy útiles para realizar un trabajo más eficiente
dentro de R
}
\author{
Esteban Rucán
}
