% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/owid_covid_data.R
\encoding{UTF-8}
\name{owid_covid_data}
\alias{owid_covid_data}
\title{Base de Datos Covid}
\usage{
owid_covid_data()
}
\value{
Tibble. Base de datos sobre el coronavirus extraida de Our World in Data
}
\description{
Función que que importa la base de datos del covid
}
\details{
Útil para no meterse a la página a cada rato.
}
\examples{
\dontrun{owid_covid_data()}

}
\author{
Esteban Rucán.
}
