% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paleta_estyr.R
\encoding{UTF-8}
\name{paleta_estyr}
\alias{paleta_estyr}
\title{Paleta EstyR}
\usage{
paleta_estyr(n = 5)
}
\arguments{
\item{n}{Valor numérico, cantidad de colores de la paleta, el máximo es 5.}
}
\value{
Paleta de colores
}
\description{
Paleta de colores
}
\details{
Está función fue desarrollada con el fin tener colores habituales
}
\examples{
\dontrun{paleta_estyr()}

}
\author{
Esteban Rucán.
}
