% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pkg_analisis.R
\encoding{UTF-8}
\name{pkg_analisis}
\alias{pkg_analisis}
\title{Cargar packages de análisis exploratorio}
\usage{
pkg_analisis()
}
\value{
Packages de análisis exploratorio cargados
}
\description{
Función que carga packages de análisis exploratorio
}
\details{
tidymodels, broom, vip, usemodels, themis
}
\examples{
\dontrun{pkg_analisis()}

}
\author{
Esteban Rucán.
}
