% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pkg_modelo.R
\encoding{UTF-8}
\name{pkg_modelos}
\alias{pkg_modelos}
\title{Cargar packages de modelamiento}
\usage{
pkg_modelos()
}
\value{
Packages de modelamiento cargados
}
\description{
Función que carga packages de modelamiento
}
\details{
Packages como tidymodels, broom, vip, usemodels, themis, etc.
}
\examples{
\dontrun{pkg_modelos()}

}
\author{
Esteban Rucán.
}
