% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pkg_series_fable.R, R/pkg_series_modeltime.R
\encoding{UTF-8}
\name{pkg_series_fable}
\alias{pkg_series_fable}
\title{[fable] Cargar packages de series de tiempo}
\usage{
pkg_series_fable()

pkg_series_fable()
}
\value{
Packages de series de tiempo cargados

Packages de series de tiempo cargados
}
\description{
Funci칩n que carga packages de series de tiempo

Funci칩n que carga packages de series de tiempo
}
\details{
ggplot2, tsibble, feasts, fable, fable.prophet

ggplot2, tsibble, feasts, fable, fable.prophet
}
\examples{
\dontrun{pkg_series_fable()}

\dontrun{pkg_series_modeltime()}

}
\author{
Esteban Ruc치n.

Esteban Ruc치n.
}
