% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pkg_series_modeltime.R
\encoding{UTF-8}
\name{pkg_series_modeltime}
\alias{pkg_series_modeltime}
\title{[modeltime] Cargar packages de series de tiempo}
\usage{
pkg_series_modeltime()
}
\value{
Packages de series de tiempo cargados
}
\description{
Función que carga packages de series de tiempo
}
\details{
ggplot2, tsibble, feasts, fable, fable.prophet
}
\examples{
\dontrun{pkg_series_modeltime()}

}
\author{
Esteban Rucán.
}
