% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme_estyr.R
\encoding{UTF-8}
\name{theme_estyr}
\alias{theme_estyr}
\title{Tema para ggplot2}
\usage{
theme_estyr(
  base_size = 11,
  strip_text_size = 12,
  strip_text_margin = 5,
  subtitle_size = 13,
  subtitle_margin = 10,
  plot_title_size = 16,
  plot_title_margin = 10,
  legend_position = "bottom",
  ...
)
}
\arguments{
\item{base_size}{Valor numérico. Tamaño de la fuente en pts.}

\item{strip_text_size}{Valor numérico. Tamaño de fuente de los títulos de las facetas.}

\item{strip_text_margin}{Valor numérico. Margen de los títulos de las facetas.}

\item{subtitle_size}{Valor numérico. Tamaño de fuente del subtítulo en pts.}

\item{subtitle_margin}{Valor numérico. Margen del subtítulo.}

\item{plot_title_size}{Valor numérico. Tamaño de fuente del título en pts.}

\item{plot_title_margin}{Valor numérico. Margen del título.}

\item{legend_position}{Character. Posición de la leyenda}
}
\value{
Plantilla de R Markdown.
}
\description{
Tema de ggplot con algunas configuraciones habituales
}
\details{
Está función fue desarrollada con el fin de tener gráficos de un solo estilo
}
\examples{
\dontrun{grafico + theme_estyr()}

}
\author{
Esteban Rucán.
}
