% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transponer.R
\encoding{UTF-8}
\name{transponer}
\alias{transponer}
\title{Transponer}
\usage{
transponer(data, ids, nombres_a = " ", prefijo = "", separador = "_")
}
\arguments{
\item{data}{Data Frame. Base de datos a transponer.}

\item{ids}{Character. Nombre de la columna de ID's}

\item{nombres_a}{Character. Nombre para la columna que tendrá los nombres de las variables de la base de datos original.}

\item{prefijo}{Character. Prefijo de las variables para la función \code{pivot_wider}.}

\item{separador}{Character. Separador de las variables antes del \code{prefijo}.}
}
\value{
Tibble. Transpuesta de un \code{Data Frame}.
}
\description{
Tranpone una base de datos de dimensión \code{n x m} a \code{m x n}.
}
\details{
Útil para voltear tibbles, ya que requiere de un proceso más largo.
}
\examples{
\dontrun{
     mtcars$row <- rownames(mtcars)
     rownames(mtcars) <- NULL
     mtcars \%>\% transponer(row)
     }

}
\author{
Esteban Rucán
}
