% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/crear_plantilla.R
\encoding{UTF-8}
\name{crear_plantilla}
\alias{crear_plantilla}
\title{Power Mean}
\usage{
crear_plantilla(
  bg = "#202123",
  fg = "#B8BCC2",
  primary = "#EA80FC",
  base_font = "Patrick Hand",
  code_font = "Fira Code",
  crear_proyecto = FALSE,
  ...
)
}
\arguments{
\item{bg}{Character. Color de fondo de la plantilla}

\item{fg}{Character. Color de fuente y lineas de la plantilla}

\item{primary}{Character. color de resaltado de la plantilla}

\item{base_font}{Character. Tipo de fuente (En Google Fonts) de la plantilla.}

\item{code_font}{Character. Tipo de fuente (En Google Fonts)  del código de la plantilla.}

\item{crear_proyecto}{Valor lógico. Si es \code{TRUE}, crea un proyecto en el directorio}

\item{...}{Parámetros adicionales para la función \code{crear_proyecto}}
}
\value{
Media generalizada según la potencia \code{m}.
}
\description{
Calcular la Media Generalizada de un vector y potencia \code{m}.
}
\details{
Las medias generalizadas, también conocidas como medias de Hölder, son una abstracción de las medias cuadráticas, aritméticas, geométricas y armónicas. Se definen y agrupan a través de la siguiente expresión:
(1 / n * sum(x ^ m)) ^ (1 / m).
}
\examples{
\dontrun{crear_plantilla()}

}
\author{
Esteban Rucán.
}
