% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/crear_plantilla.R
\encoding{UTF-8}
\name{crear_plantilla}
\alias{crear_plantilla}
\title{Crear Plantilla de R Markdown.}
\usage{
crear_plantilla(
  titulo,
  color_fondo = "#202123",
  color_fuente = "#B8BCC2",
  resaltado = "#EA80FC",
  tipo_fuente = "Patrick Hand",
  tipo_fuente_codigo = "Fira Code"
)
}
\arguments{
\item{titulo}{Character. Título de la plantilla.}

\item{color_fondo}{Character. Color de fondo de la plantilla}

\item{color_fuente}{Character. Color de fuente y lineas de la plantilla}

\item{resaltado}{Character. color de resaltado de la plantilla}

\item{tipo_fuente}{Character. Tipo de fuente (En Google Fonts) de la plantilla.}

\item{tipo_fuente_codigo}{Character. Tipo de fuente (En Google Fonts) del código de la plantilla.}
}
\value{
Plantilla de R Markdown.
}
\description{
Crear una plantilla de R Markdown en el directorio seleccionado.
}
\details{
Esta función proporciona una forma rápida de crear una plantilla de R Markdown.
}
\examples{
\dontrun{crear_plantilla("Analsis Exploratorio")}

}
\author{
Esteban Rucám.
}
